fix: resolve clippy warnings across Rust workspace
- Allow needless_range_loop where index serves dual purpose
- Allow too_many_arguments on SIMD process_chunk functions
- Remove unnecessary unsafe block in AVX-512 quote filter
- Use copy_from_slice for clean LUT initialization
- Use Iterator::find for literal allocation
- Remove unnecessary cast in bench data generator

+++ rust/findchars-bench/src/lib.rs
@@ -65,7 +65,7 @@ pub fn generate_csv_data(
             if quoted {
                 data.push(b'"');
                 for _ in 0..field_len {
-                    let ch = (b'a' + rng.random_range(0..26u8)) as u8;
+                    let ch = b'a' + rng.random_range(0..26u8);
                     data.push(ch);
                     // Occasionally embed a comma or quote
                     if rng.random_range(0..20) == 0 {
